standardise solution structure
