+++ StardewValleyMods.sln
@@ -1,0 +1,25 @@
+﻿
+Microsoft Visual Studio Solution File, Format Version 12.00
+# Visual Studio 15
+VisualStudioVersion = 15.0.27004.2002
+MinimumVisualStudioVersion = 10.0.40219.1
+Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "24hClock", "24hClock\24hClock.csproj", "{CA0570BB-53A4-45DC-9E49-18AD7A419582}"
+EndProject
+Global
+	GlobalSection(SolutionConfigurationPlatforms) = preSolution
+		Debug|Any CPU = Debug|Any CPU
+		Release|Any CPU = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(ProjectConfigurationPlatforms) = postSolution
+		{CA0570BB-53A4-45DC-9E49-18AD7A419582}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
+		{CA0570BB-53A4-45DC-9E49-18AD7A419582}.Debug|Any CPU.Build.0 = Debug|Any CPU
+		{CA0570BB-53A4-45DC-9E49-18AD7A419582}.Release|Any CPU.ActiveCfg = Release|Any CPU
+		{CA0570BB-53A4-45DC-9E49-18AD7A419582}.Release|Any CPU.Build.0 = Release|Any CPU
+	EndGlobalSection
+	GlobalSection(SolutionProperties) = preSolution
+		HideSolutionNode = FALSE
+	EndGlobalSection
+	GlobalSection(ExtensibilityGlobals) = postSolution
+		SolutionGuid = {79BCBBCB-325E-4647-9BB7-841F0B54949F}
+	EndGlobalSection
+EndGlobal
