--- manifest.json
@@ -1,9 +1,0 @@
-{
-   "Name": "24h Clock",
-   "Author": "Lajna",
-   "Version": "0.0.1",
-   "Description": "Makes the clock show time in 24h format.",
-   "UniqueID": "Lajna.24hClock",
-   "EntryDll": "24hClock.dll",
-   "UpdateKeys": []
-}
